fix: get h264 direct working
(only works after delay due to waiting for IDR frame)
