+++ src/ros/cameras/cameras2++/include/cameras/cameras.hpp
@@ -1,5 +1,4 @@
 #include "rclcpp/rclcpp.hpp"
-#include "cameras/pipeline.hpp"
 #include <camera_msgs/msg/camera.hpp>
 #include <gst/gst.h>
 
@@ -11,23 +10,6 @@
 #define SERVICE_DISCOVERY   "/camera_directory/discover"
 #define TOPIC_CAMERAS       "/camera_directory/cameras"
 #define POLLING_PERIOD      5000
-#define PIPELINE_PREFIX     "serial_pipelines"
 
 extern rclcpp::QoS discover_qos; 
-
-struct Pipeline
-{
-  GstElement* gst_pipeline;
-  Properties* props;
-  std::string pipeline_type;
-  camera_msgs::msg::Camera* camera;
-};
-
-struct v4l2webrtcPipelineProperties : Properties, v4lProperties, webRTCProperties, clockProperties {};
-v4l2webrtcPipelineProperties* get_v4l2webrtc_pipeline_properties(rclcpp::Node* log_node, camera_msgs::msg::Camera* camera);
-GstElement* v4l2webrtc_pipeline(rclcpp::Node* log_node, v4l2webrtcPipelineProperties* props);
-
-struct h264directPipelineProperties : Properties, v4lProperties, webRTCProperties, clockProperties {};
-h264directPipelineProperties* get_h264direct_pipeline_properties(rclcpp::Node* log_node, camera_msgs::msg::Camera* camera);
-GstElement* h264direct_pipeline(rclcpp::Node* log_node, h264directPipelineProperties* props);
